# core/spec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cuate API",
    "version": "1",
    "description": "The one owner of the contract between the clients and the server: every route, scope, event and payload. The server's routes and the client's calls are both tested against it (server/test/api.test.js). A payload key not declared here fails the test."
  },
  "paths": {
...
    "/api/v1/chats/{chatId}/messages": {
      "get": {
        "operationId": "messages",
        "summary": "GET /api/v1/chats/:chatId/messages",
        "tags": [
          "api"
        ],
        "x-scope": "read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageList"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "send",
        "summary": "POST /api/v1/chats/:chatId/messages",
        "tags": [
          "api"
        ],
        "x-scope": "send",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendResult"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRequest"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "operationId": "settings",
        "summary": "GET /api/v1/settings",
        "tags": [
          "api"
        ],
        "x-scope": "read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "settings-write",
        "summary": "PUT /api/v1/settings",
        "tags": [
          "api"
        ],
        "x-scope": "settings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "default": {
            "description": "An error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "MessageList": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "messages",
          "hasMore"
        ],
        "additionalProperties": false
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chatId": {
            "type": "string"
          },
          "fromMe": {
            "type": "boolean"
          },
          "sender": {
            "type": "string"
          },
          "senderName": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "sentAt": {
            "type": "string"
          },
          "replyTo": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reaction"
            }
          }
        },
        "required": [
          "id",
          "chatId",
          "fromMe",
          "text",
          "sentAt",
          "attachments",
          "reactions"
        ],
        "additionalProperties": false
      },
      "Attachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "mime": {
            "type": "string"
          },
          "bytes": {
            "type": "number"
          },
          "sticker": {
            "type": "boolean"
          },
          "missing": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "mime",
          "bytes",
          "sticker",
          "missing"
        ],
        "additionalProperties": false
      },
      "Reaction": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "emoji": {
            "type": "string"
          },
          "fromMe": {
            "type": "boolean"
          },
          "sender": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "fromMe"
        ],
        "additionalProperties": false
      },
      "SendRequest": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "file": {
            "type": "string"
          },
          "clientKey": {
            "type": "string"
          }
        },
        "required": [
          "clientKey"
        ],
        "additionalProperties": false
      },
...
      "SendResult": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "clientKey": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "duplicate": {
            "type": "boolean"
          }
        },
        "required": [
          "status",
          "clientKey"
        ],
        "additionalProperties": false
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "values": {
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "additionalProperties": false
      },
      "SettingsPatch": {
        "type": "object",
        "properties": {
          "values": {
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}